requirements.txt
argon2-cffi==18.3.0
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
Django==3.1.5
django-anymail==8.2
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
httpie==2.3.0
idna==2.10
Markdown==3.3.3
mccabe==0.6.1
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.4
PyJWT==2.0.1
PySocks==1.7.1
pytz==2020.5
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
